+++ components/GeoTriggering.js
@@ -1,20 +1,49 @@
 import React, { useState, useEffect } from "react";
-import { Button, Platform, Text, View } from "react-native";
+import { Button, Platform, Text, View, Switch } from "react-native";
 import { useNavigate } from "react-router";
 import BluedotPointSdk from "bluedot-react-native";
 import styles from "../styles";
+import AsyncStorage from "@react-native-async-storage/async-storage";
+import { OS } from "../enums";
 
 export default function GeoTriggering() {
+
   const navigate = useNavigate();
   const [isGeotriggeringRunning, setIsGeotriggeringRunning] = useState(false);
   const [error, setError] = useState(null);
+  const [isBackgroundLocationUpdatesEnabled, setIsBackgroundLocationUpdatesEnabled] = useState(false);
+  const allowsBackgroundLocationUpdatesString = "allowsBackgroundLocationUpdates";
+  const toggleSwitch = async (value) => {
+    setIsBackgroundLocationUpdatesEnabled(value);
+    BluedotPointSdk.allowsBackgroundLocationUpdates(value);
+    try {
+    await AsyncStorage.setItem(allowsBackgroundLocationUpdatesString, value.toString());
+    } catch (e) {
+      console.log(e);
+    }
+  };
 
   const geoTriggeringBuilder = new BluedotPointSdk.GeoTriggeringBuilder();
+  let isBackgroundLocationUpdatesEnabledString = `Is Background Location Enabled: ${isBackgroundLocationUpdatesEnabled}`;
 
   useEffect(() => {
     BluedotPointSdk.isGeoTriggeringRunning().then((isRunning) => {
       setIsGeotriggeringRunning(isRunning);
     });
+
+    const retrieveBackgroundLocationStatus = async () => {
+      try {
+        const isBackgroundLocationUpdatesAllowed = ((await AsyncStorage.getItem(allowsBackgroundLocationUpdatesString) || 'false') === 'true')
+        if(isBackgroundLocationUpdatesAllowed !== null) {
+          setIsBackgroundLocationUpdatesEnabled(isBackgroundLocationUpdatesAllowed);
+        }
+      } catch(e) {
+        console.log(e);
+      }
+    };
+    
+    retrieveBackgroundLocationStatus();
+
   }, []);
 
   const handleStartGeotriggering = () => {
@@ -64,7 +93,7 @@ export default function GeoTriggering() {
   const renderStartButtons = () => {
     return (
       <>
-        {Platform.OS === "android" && (
+        {Platform.OS === OS.ANDROID && (
           <Button
             title={"Start with foreground notification"}
             onPress={handleStartGeotriggeringWithAndroidNotification}
@@ -85,6 +114,14 @@ export default function GeoTriggering() {
     <View style={styles.container}>
       <Text style={styles.eventTitle}>Geo Triggering</Text>
       {error ? <Text>Error: {error}</Text> : null}
+      {Platform.OS === OS.IOS && <>
+        <Text>Allow Background Location Updates</Text>
+        <Switch
+          onValueChange={toggleSwitch}
+          value={isBackgroundLocationUpdatesEnabled}
+        />
+        <Text>{isBackgroundLocationUpdatesEnabledString}</Text>
+      </>}
       {isGeotriggeringRunning ? (
         <Button
           title={"Stop"}
@@ -94,7 +131,6 @@ export default function GeoTriggering() {
       ) : (
         renderStartButtons()
       )}
-
       <Button
         title="Back"
         onPress={() => navigate("/main")}