+++ ReleaseNotes.md
@@ -1,4 +1,4 @@
 # Bluedot ReactNative Minimal Integration
 
-- Using [Bluedot plugin 2.2.0](https://github.com/Bluedot-Innovation/Bluedot-React-Native-Plugin) (support react-native@0.69.4)
-- Recreated the Minimal Integration App with react@18.0.0, react-native@0.69.4
+- Update UI to enable/disable background location updates on iOS when the app is in While in use location permission.
+- Using [Bluedot plugin 2.3.0](https://github.com/Bluedot-Innovation/Bluedot-React-Native-Plugin) (support react-native@>=0.69.0).